Added confirmation to the ExpenseEditor
