+++ .changeset/moody-tomatoes-try.md
@@ -1,0 +1,6 @@
+---
+"@inkbeard/budget-it": minor
+---
+
+- Added `ConfirmationService` installation to the plugin installation.
+- Added confirmation of expense deletion
