+++ .changeset/soft-buckets-chew.md
@@ -1,0 +1,6 @@
+---
+"@inkbeard/budget-it": minor
+---
+
+- Added ExpenseEditor component
+- Replaced expense form in ExpenseItem component with ExpenseEditor
