+++ .changeset/swift-spiders-pump.md
@@ -1,0 +1,5 @@
+---
+"@inkbeard/ui-library": minor
+---
+
+Fixed documetation for the AppConfirmPopup component
