+++ .changeset/warm-brooms-wave.md
@@ -1,0 +1,5 @@
+---
+"inkbeard": minor
+---
+
+installed Budget It plugin
